updatehub-active-set: Support override of env and fw_env.config
We eventually need to do a more convoluted update process which involves
adding redundant environment and/or changing the default
environment.

For that, we need to use the fw_printenv in a more advanced way and use
local files which are extrated from the update package itself. It uses
two hard coded places for them:

 - fw_env.config -> /tmp/updatehub.fw_env_config
 - defenv -> /tmp/updatehub.defenv

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>